refactor(region-builder): slim parent into hooks + section components
Splits RegionBuilder.tsx (1089 LoC) into a 193 LoC orchestrator that
composes the previously-extracted hooks with new presentational sections
under region-builder/sections/:

- HealthBanner            partial-deploy alert
- ActiveJobsTable         provisioning-in-flight rows + steps + diag drawer
- ProvisionedRegionsTable per-profile graph readiness + OverflowMenu actions
- BuildHistoryTable       ORS_BUILD_HISTORY rerun table
- ProvisionForm           source tab + catalog list + profiles + size + PBF
- FailedJobsList          error cards with retry / dismiss + diag drawer
- CompletedJobsTable      recent completed jobs

JSX, polling intervals (3s jobs / 5s build progress), and the auto-refresh
catalog flow are preserved verbatim. The diag drawer state is shared via
a single useDiagnostics hook so Active and Failed sections stay in sync.

Verified with `npx tsc --noEmit` and `npm run build`.
